dom/workers/test/test_sharedWorker_ports.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
  <head>
    <title>Test for MessagePort and SharedWorkers</title>
    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
  <body>
    <script class="testbody" type="text/javascript">

var sw1 = new SharedWorker('sharedWorker_ports.js');
sw1.port.onmessage = function(event) {
  if (event.data.type == "connected") {
    ok(true, "The SharedWorker is alive.");

    var sw2 = new SharedWorker('sharedWorker_ports.js');
    sw1.port.postMessage("Port from the main-thread!", [sw2.port]);
    return;
  }

  if (event.data.type == "status") {
    ok(event.data.test, event.data.msg);
    return;
  }

  if (event.data.type == "finish") {
    info("Finished!");
    ok(sw1.port, "The port still exists");
    sw1.port.foo = sw1; // Just a test to see if we leak.
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();
    </script>
  </body>
</html>
